requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
multidict==6.1.0
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
schedule==1.2.2
SQLAlchemy==2.0.35
Telethon==1.36.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
